[Fix] 검색 시, 최근 검색어에 추가되도록 반영, 입력창 분리, 컴포넌트 출력 시 분기 처리 수정 #25
