+++ src/components/header/use-header.ts
@@ -1,16 +1,13 @@
 import { useEffect, useState } from "react";
 import { throttle } from "lodash-es";
-import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { usePathname, useRouter } from "next/navigation";
 
 export default function useHeader() {
   // const [isSidebarOpen, setIsSidebarOpen] = useState<boolean>(false);
   const currentPath = usePathname();
   const isMainPath =
     currentPath === "/main" || currentPath === "/" ? false : true;
-  const searchParams = useSearchParams();
-  const currentKeyword = searchParams?.get("keyword");
 
-  const [keyword, setKeyword] = useState<string>(currentKeyword ?? "");
   const [shadow, setShadow] = useState(isMainPath);
   const [openSidebar, setOpenSidebar] = useState<boolean>(false);
 
@@ -59,8 +56,6 @@ export default function useHeader() {
     shadow,
     router,
     currentPath,
-    keyword,
-    setKeyword,
     openSidebar,
     setOpenSidebar,
   };
